+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Get control characters from termios for emulating SIGINT/SIGQUIT/EOF.
+
 ## [0.4.0] - 2025-05-18
 
 ### Changed